anread-ocr/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==11.1.0
httpx==0.28.1